requirements.txt
# Основные зависимости
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
python-dotenv==1.0.1

# OpenAI и AI
openai==1.57.4

# Обработка данных (версии для Python 3.9)
pandas==2.0.3
numpy==1.26.4
openpyxl==3.1.5

# Обработка документов
PyPDF2==3.0.1
pdf2image==1.17.0
Pillow==11.0.0
python-docx==1.1.2

# RAG и векторный поиск
faiss-cpu==1.9.0.post1
langchain-text-splitters==0.3.4

# Логирование и мониторинг
tqdm==4.67.1

# Для knowledge_base_builder (парсинг статей и законов)
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Валидация данных
pydantic==2.10.4
pydantic-core==2.27.2

# ASGI сервер
anyio==4.7.0

